resume-parser/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
sqlalchemy>=2.0.0
pymupdf>=1.24.0
python-docx>=1.1.0
python-dotenv>=1.0.0
openai>=1.12.0
spacy>=3.7.0
pandas>=2.2.0
pydantic>=2.5.0
